card reader manager added
