+++ congateway/pca10040/s132/armgcc/Makefile
@@ -69,6 +69,7 @@ SRC_FILES += \
   $(PROJ_DIR)/ble_fsm.c \
   $(PROJ_DIR)/ble_fsm_actions.c \
   $(PROJ_DIR)/software_uart.c \
+  $(PROJ_DIR)/card_reader_manager.c \
   $(SDK_ROOT)/external/segger_rtt/SEGGER_RTT.c \
   $(SDK_ROOT)/external/segger_rtt/SEGGER_RTT_Syscalls_GCC.c \
   $(SDK_ROOT)/external/segger_rtt/SEGGER_RTT_printf.c \
